+++ blosc/context.h
@@ -129,7 +129,7 @@ struct thread_context {
   uint8_t* tmp3;
   uint8_t* tmp4;
   int32_t tmp_blocksize; /* the blocksize for different temporaries */
-  int32_t tmp_nbytes;   /* keep track of how big the temporary buffers are */
+  size_t tmp_nbytes;   /* keep track of how big the temporary buffers are */
 #if defined(HAVE_ZSTD)
   /* The contexts for ZSTD */
   ZSTD_CCtx* zstd_cctx;
